+++ datadog/resource_datadog_timeboard.go
@@ -41,6 +41,13 @@ func resourceDatadogTimeboard() *schema.Resource {
 					Type:     schema.TypeMap,
 					Optional: true,
 				},
+				"metadata_json": {
+					Type:     schema.TypeString,
+					Optional: true,
+					// NOTE: this is using functions from resource_datadog_screenboard
+					// since the metadata attribute is the same for both of these boards
+					ValidateFunc: validateMetadataJSON,
+				},
 				"conditional_format": {
 					Type:        schema.TypeList,
 					Optional:    true,
@@ -104,6 +111,13 @@ func resourceDatadogTimeboard() *schema.Resource {
 					Type:        schema.TypeString,
 					Optional:    true,
 					Description: "If set to 'present', this will include the present values in change graphs.",
+					ValidateFunc: func(val interface{}, key string) (warns []string, errs []error) {
+						stringVal := val.(string)
+						if stringVal != "" && stringVal != "present" {
+							errs = append(errs, fmt.Errorf("'%s' value must be empty or 'present', got: '%s'", key, stringVal))
+						}
+						return
+					},
 				},
 			},
 		},
@@ -190,6 +204,14 @@ func resourceDatadogTimeboard() *schema.Resource {
 					Type:        schema.TypeString,
 					Optional:    true,
 					Description: "How many digits to show",
+					// Suppress the diff shown if the graph is going to be set to a default of 2 iff its not set by the config
+					// Since this precision attribute is only valid for certain graph types, we aren't setting a global default
+					DiffSuppressFunc: func(k, old, new string, d *schema.ResourceData) bool {
+						if old == "2" && new == "" {
+							return true
+						}
+						return false
+					},
 				},
 				"custom_unit": {
 					Type:        schema.TypeString,
@@ -349,7 +371,7 @@ func buildTemplateVariables(terraformTemplateVariables *[]interface{}) *[]datado
 	return &datadogTemplateVariables
 }
 
-func appendRequests(datadogGraph *datadog.Graph, terraformRequests *[]interface{}) {
+func appendRequests(datadogGraph *datadog.Graph, terraformRequests *[]interface{}) error {
 	for _, _t := range *terraformRequests {
 		t := _t.(map[string]interface{})
 		log.Printf("[DataDog] request: %v", pretty.Sprint(t))
@@ -394,7 +416,12 @@ func appendRequests(datadogGraph *datadog.Graph, terraformRequests *[]interface{
 			d.SetOrderBy(v.(string))
 		}
 		if v, ok := t["extra_col"]; ok {
-			d.SetExtraCol(v.(string))
+			// additional validation: `extra_col` may only be used for `change` viz
+			if viz := datadogGraph.Definition.GetViz(); viz == "change" {
+				d.SetExtraCol(v.(string))
+			} else if v != nil && v != "" {
+				return fmt.Errorf("'extra_col' attribute may only be used for 'change' viz, not '%s'", viz)
+			}
 		}
 		if v, ok := t["order_direction"]; ok {
 			d.SetOrderDirection(v.(string))
@@ -404,9 +431,15 @@ func appendRequests(datadogGraph *datadog.Graph, terraformRequests *[]interface{
 			_v := v.([]interface{})
 			appendConditionalFormats(&d, &_v)
 		}
+		if v, ok := t["metadata_json"]; ok {
+			d.Metadata = map[string]datadog.GraphDefinitionMetadata{}
+			getMetadataFromJSON([]byte(v.(string)), &d.Metadata)
+		}
 
 		datadogGraph.Definition.Requests = append(datadogGraph.Definition.Requests, d)
 	}
+
+	return nil
 }
 
 func appendEvents(datadogGraph *datadog.Graph, terraformEvents *[]interface{}) {
@@ -431,7 +464,7 @@ func appendMarkers(datadogGraph *datadog.Graph, terraformMarkers *[]interface{})
 	}
 }
 
-func buildGraphs(terraformGraphs *[]interface{}) *[]datadog.Graph {
+func buildGraphs(terraformGraphs *[]interface{}) (*[]datadog.Graph, error) {
 	datadogGraphs := make([]datadog.Graph, len(*terraformGraphs))
 	for i, _t := range *terraformGraphs {
 		t := _t.(map[string]interface{})
@@ -543,9 +576,12 @@ func buildGraphs(terraformGraphs *[]interface{}) *[]datadog.Graph {
 		appendEvents(d, &v)
 
 		v = t["request"].([]interface{})
-		appendRequests(d, &v)
+		err := appendRequests(d, &v)
+		if err != nil {
+			return nil, err
+		}
 	}
-	return &datadogGraphs
+	return &datadogGraphs, nil
 }
 
 func buildTimeboard(d *schema.ResourceData) (*datadog.Dashboard, error) {
@@ -559,12 +595,16 @@ func buildTimeboard(d *schema.ResourceData) (*datadog.Dashboard, error) {
 	}
 	terraformGraphs := d.Get("graph").([]interface{})
 	terraformTemplateVariables := d.Get("template_variable").([]interface{})
+	graphs, err := buildGraphs(&terraformGraphs)
+	if err != nil {
+		return nil, err
+	}
 	return &datadog.Dashboard{
 		Id:                datadog.Int(id),
 		Title:             datadog.String(d.Get("title").(string)),
 		Description:       datadog.String(d.Get("description").(string)),
 		ReadOnly:          datadog.Bool(d.Get("read_only").(bool)),
-		Graphs:            *buildGraphs(&terraformGraphs),
+		Graphs:            *graphs,
 		TemplateVariables: *buildTemplateVariables(&terraformTemplateVariables),
 	}, nil
 }
@@ -579,7 +619,7 @@ func resourceDatadogTimeboardCreate(d *schema.ResourceData, meta interface{}) er
 		return fmt.Errorf("Failed to create timeboard using Datadog API: %s", err.Error())
 	}
 	d.SetId(strconv.Itoa(timeboard.GetId()))
-	return nil
+	return resourceDatadogTimeboardRead(d, meta)
 }
 
 func resourceDatadogTimeboardUpdate(d *schema.ResourceData, meta interface{}) error {
@@ -653,6 +693,10 @@ func appendTerraformGraphRequests(datadogRequests []datadog.GraphDefinitionReque
 		}
 		if v, ok := datadogRequest.GetExtraColOk(); ok {
 			request["extra_col"] = v
+		}
+		if datadogRequest.Metadata != nil {
+			res, _ := json.Marshal(datadogRequest.Metadata)
+			request["metadata_json"] = string(res)
 		}
 
 		*requests = append(*requests, request)
@@ -811,7 +855,6 @@ func resourceDatadogTimeboardRead(d *schema.ResourceData, meta interface{}) erro
 	// Ensure the ID saved in the state is always the legacy ID returned from the API
 	// and not the ID passed to the import statement which could be in the new ID format
 	d.SetId(strconv.Itoa(timeboard.GetId()))
-
 	return nil
 }